refactor: project entity
