+++ packages/backend/src/app/helper/telemetry.utils.ts
@@ -2,7 +2,7 @@ import { PostHog } from 'posthog-node'
 import { SystemProp } from './system/system-prop'
 import { system } from './system/system'
 import { ProjectId, TelemetryEvent, User, UserId } from '@activepieces/shared'
-import { projectService } from '../project/project.service'
+import { projectService } from '../project/project-service'
 import { getEdition } from './secret-helper'
 import { logger } from './logger'
